End-to-End-medical-Chatbot-main/requirements.txt
sentence-transformers==2.2.2
langchain==0.0.354
flask==2.2.5
PyPDF2==3.0.1
python-dotenv==0.21.1
pinecone-client==2.2.4
transformers==4.21.3
accelerate==0.21.0
torch==1.13.1
numpy==1.21.6
scikit-learn==1.0.2
requests==2.28.2